pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = ""
authors = [
    {name = "dvilela",email = "dvilelaf@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "customtkinter (>=5.2.2,<6.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
